Moved Event attributes
